LayoutTests/fast/dom/DOMURL/searchparams-expected.txt

CONSOLE MESSAGE: line 8: PASS
CONSOLE MESSAGE: line 11: PASS
